Refactor score function to be in utils
