+++ rec/metrics/__init__.py
@@ -1,6 +1,7 @@
 """ Export various measurements that users can plug into their simulations """
 from .measurement import (
     Measurement,
+    MeasurementModule,
     HomogeneityMeasurement,
     JaccardSimilarity,
     MSEMeasurement,
